+++ src/mt19937-64.c
@@ -149,19 +149,19 @@ long long genrand64_int63(void)
 }
 
 /* generates a random number on [0,1]-real-interval */
-float genrand64_real1(void)
+double genrand64_real1(void)
 {
     return (genrand64_int64() >> 11) * (1.0/9007199254740991.0);
 }
 
 /* generates a random number on [0,1)-real-interval */
-float genrand64_real2(void)
+double genrand64_real2(void)
 {
     return (genrand64_int64() >> 11) * (1.0/9007199254740992.0);
 }
 
 /* generates a random number on (0,1)-real-interval */
-float genrand64_real3(void)
+double genrand64_real3(void)
 {
     return ((genrand64_int64() >> 12) + 0.5) * (1.0/4503599627370496.0);
 }
