+++ src/tests/test_vortexMultiRun.c
@@ -14,16 +14,16 @@ int main(int argc,char **argv){
   int seed=98755;
   int i,j,err,ngbr,found,nCnect,*label,n,bin,nMax=20;
   int nbList[8],eqList[Pop],**eqClass;
-  float Gmin=1,Gmax=2,rmin=1,rmax=1;
-  float xmin[2]={-9.,-9.},xmax[2]={9.,9.};
-  float *parVortex=NULL,x0[2],dx[2],xf[2],*sField=NULL,*gField;
-  float x,y,v0y0 = 0.00,*vCatalog=NULL;
+  double Gmin=1,Gmax=2,rmin=1,rmax=1;
+  double xmin[2]={-9.,-9.},xmax[2]={9.,9.};
+  double *parVortex=NULL,x0[2],dx[2],xf[2],*sField=NULL,*gField;
+  double x,y,v0y0 = 0.00,*vCatalog=NULL;
   FILE *dadosgen,*dadosref,*dadosout;
   
   x0[0]=-10.; xf[0]= 10.; dx[0] = (xf[0]-x0[0])/Height;
   x0[1]=-10.; xf[1]= 10.; dx[1] = (xf[1]-x0[1])/Width;
   
-  gField = (float *)malloc(4*Height*Width*sizeof(float));
+  gField = (double *)malloc(4*Height*Width*sizeof(double));
   if(gField==NULL){
     printf("memory not allocked\n");
     return 1;
@@ -35,7 +35,7 @@ int main(int argc,char **argv){
     return 2;
   }
 
-  vCatalog = (float*)malloc(4*nMax*sizeof(float));
+  vCatalog = (double*)malloc(4*nMax*sizeof(double));
   if(vCatalog==NULL){
     printf("memory not allocked\n");
     return 3;
